move tests into separate files; mark consts in rp2_dma const
